Add more combat profile values

+++ nitrogen-api/src/main/java/me/biquaternions/nitrogen/entity/CombatProfile.java
@@ -22,6 +22,6 @@ public abstract class CombatProfile {
 
     public abstract float getRegenerationExhaustion();
 
-    public abstract double getCriticalMultiplier();
+    public abstract float getCriticalMultiplier();
 
 }
